+++ src/components/CConnectButton/modal.tsx
@@ -2,8 +2,9 @@ import React, { useEffect, useRef } from 'react';
 import Link from 'next/link';
 import Image from 'next/image';
 
+import toast from '../CToast';
 import CButton from '../CButton';
-import CModalSheet from '../CModalSheet';
+import CBottomSheet from '../CBottomSheet';
 
 import copy from 'public/images/copy.svg';
 import power from 'public/images/power.svg';
@@ -22,9 +23,12 @@ type ModalProps = {
   address: string;
   isMinimized: boolean;
   closeModal: () => void;
+  setIsOpen: (_: boolean) => void;
 };
-const Modal = ({ open, address, closeModal, isMinimized }: ModalProps) => {
+const Modal = ({ open, address, closeModal, isMinimized, setIsOpen }: ModalProps) => {
   const modalRef = useRef<HTMLDivElement | null>(null);
+  const bottomSheetRef = useRef<HTMLDivElement | null>(null);
+
   const dispatch = useAppDispatch();
 
   const handleDisconnect = () => {
@@ -32,13 +36,20 @@ const Modal = ({ open, address, closeModal, isMinimized }: ModalProps) => {
     dispatch(clearTokenBalances());
   };
 
-  const handleCopy = () => {
+  const handleCopy = (e: React.MouseEvent<HTMLImageElement, MouseEvent>) => {
+    e.stopPropagation(); // Stop event propagation
     copyText(address);
+    toast('success', 'Address copied successfully');
   };
 
   useEffect(() => {
     const handleClickOutside = (event: MouseEvent) => {
-      if (modalRef.current && !modalRef.current.contains(event.target as Node)) {
+      if (
+        modalRef.current &&
+        !modalRef.current.contains(event.target as Node) &&
+        bottomSheetRef.current &&
+        !bottomSheetRef.current.contains(event.target as Node)
+      ) {
         closeModal();
       }
     };
@@ -58,15 +69,18 @@ const Modal = ({ open, address, closeModal, isMinimized }: ModalProps) => {
     <>
       <div
         ref={modalRef}
-        className={`bg-midnightBlue p-[6px] rounded-[10px] mobile:hidden${
+        className={`bg-midnightBlue p-[6px] rounded-[10px] mobile:!hidden ${
           open
             ? `fixed bottom-11 ${isMinimized ? 'left-[100px]' : 'left-[20%]'} w-[203px] z-50`
             : 'hidden'
         }`}
       >
-        <div className="h-12 px-2 rounded-md bg-white text-midnightBlue flex justify-between items-center w-full">
+        <div
+          className="h-12 px-2 rounded-md bg-white text-midnightBlue flex justify-between items-center w-full"
+          onClick={handleCopy}
+        >
           <span>{shortenAddress(address, 5)}</span>
-          <Image src={copy} alt="copy" onClick={handleCopy} />
+          <Image src={copy} alt="copy" />
         </div>
         <div className="flex flex-col divide-y divide-slate-700 mt-2 px-2 text-white">
           <div className="py-2 flex justify-between items-center w-full">
@@ -85,13 +99,21 @@ const Modal = ({ open, address, closeModal, isMinimized }: ModalProps) => {
           </div>
         </div>
       </div>
-      <CModalSheet open={open} onClose={closeModal} contentClass="!px-6">
-        <div ref={modalRef}>
+      <CBottomSheet
+        isOpen={open}
+        setIsOpen={setIsOpen}
+        contentClass="!px-6"
+        className="desktop:!hidden !z-[9998]"
+      >
+        <div ref={bottomSheetRef}>
           <p className="text-2xl font-medium text-midnightBlue">Wallet</p>
           <div className="flex flex-col items-center justify-center gap-2 mt-6">
-            <div className="h-12 px-4 rounded-md bg-alabaster text-midnightBlue flex justify-between items-center w-full">
+            <div
+              onClick={handleCopy}
+              className="h-12 px-4 rounded-md bg-alabaster text-midnightBlue flex justify-between items-center w-full"
+            >
               <span>{shortenAddress(address, 5)}</span>
-              <Image src={copy} alt="copy" onClick={handleCopy} />
+              <Image src={copy} alt="copy" />
             </div>
             <Link
               className="flex mt-1 justify-center h-12 items-center gap-2 !text-royalBlue w-full border rounded-[10px] border-royalBlue"
@@ -110,7 +132,7 @@ const Modal = ({ open, address, closeModal, isMinimized }: ModalProps) => {
             />
           </div>
         </div>
-      </CModalSheet>
+      </CBottomSheet>
     </>
   );
 };
